Fixed a zero size/pos when resizing is disabled
The window only sets it's dimensions on resize. The window creation actually sends the events, but because we have to register the callbacks AFTER window creation, the first events are obviously missed. For now we manually set those variables after creation, might have to see if there is a way to automate this.